+++ lib/app/data/models/debug_model.dart
@@ -1,0 +1,1 @@
+enum LogLevel { error, warning, info }
