+++ lib/app/modules/debug/bindings/debug_binding.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+import '../controllers/debug_controller.dart';
+
+class DebugBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<DebugController>(
+      () => DebugController(),
+    );
+  }
+}
