+++ lib/app/routes/app_routes.dart
@@ -15,6 +15,7 @@ abstract class Routes {
   static const STOPWATCH = _Paths.STOPWATCH;
   static const ADD_OR_UPDATE_PROFILE = _Paths.ADD_OR_UPDATE_PROFILE;
   static const NOTIFICATIONS = _Paths.NOTIFICATIONS;
+  static const DEBUG = _Paths.DEBUG;
 }
 
 abstract class _Paths {
@@ -31,4 +32,5 @@ abstract class _Paths {
   static const STOPWATCH = '/stopwatch';
   static const ADD_OR_UPDATE_PROFILE = '/add-or-update-profile';
   static const NOTIFICATIONS = '/notifications';
+  static const DEBUG = '/debug';
 }
