+++ lib/app/utils/end_drawer.dart
@@ -4,7 +4,9 @@ import 'package:ultimate_alarm_clock/app/modules/settings/controllers/theme_cont
 import 'package:ultimate_alarm_clock/app/routes/app_pages.dart';
 import 'package:ultimate_alarm_clock/app/utils/constants.dart';
 import 'package:ultimate_alarm_clock/app/utils/utils.dart';
-import 'package:ultimate_alarm_clock/app/modules/debug/views/debug_screen.dart';
+import 'package:ultimate_alarm_clock/app/modules/debug/views/debug_view.dart';
+import 'package:ultimate_alarm_clock/app/modules/debug/views/debug_view.dart';
+import 'package:ultimate_alarm_clock/app/modules/debug/controllers/debug_controller.dart';
 
 Widget buildEndDrawer(BuildContext context) {
   ThemeController themeController = Get.find<ThemeController>();
@@ -81,7 +83,7 @@ Widget buildEndDrawer(BuildContext context) {
             onTap: () {
               Utils.hapticFeedback();
               Get.back();
-              Get.to(() => const DebugScreen());
+              Get.toNamed(Routes.DEBUG);
             },
             contentPadding: const EdgeInsets.only(left: 20, right: 44),
             title: Text(
